+++ src/domain/profile/hooks/useProfile.ts
@@ -2,6 +2,8 @@ import { useQuery } from '@apollo/react-hooks';
 import { useDispatch, useSelector } from 'react-redux';
 import * as Sentry from '@sentry/browser';
 import { QueryResult as GenericQueryResult } from '@apollo/react-common';
+import { useHistory } from 'react-router';
+import { useTranslation } from 'react-i18next';
 
 import { isAuthenticatedSelector } from '../../auth/state/AuthenticationSelectors';
 import {
@@ -21,6 +23,8 @@ export type ProfileQueryResult = Omit<
 };
 
 function useProfile(): ProfileQueryResult {
+  const { i18n } = useTranslation();
+  const history = useHistory();
   const dispatch = useDispatch();
   const isAuthenticated = useSelector(isAuthenticatedSelector);
 
@@ -33,6 +37,18 @@ function useProfile(): ProfileQueryResult {
       dispatch(saveProfile(data?.myProfile || defaultProfileData));
       dispatch(clearEvent());
       dispatch(saveChildrenEvents(data?.myProfile?.children || undefined));
+
+      // If the user has no profile it means that they have not yet
+      // registered to kukkuu. In this case we want to redirect them
+      // into the landing page where they can start the registration
+      // process.
+
+      // This query should be skipped when the user is not
+      // authenticated. However, it seems that this does not always
+      // hold and we have to check authentication again in the callback.
+      if (isAuthenticated && !data?.myProfile) {
+        history.replace(`/${i18n.language}/home#register`);
+      }
     },
     onError: (error) => {
       // eslint-disable-next-line no-console